+++ packages/jacaranda/src/runtime.js
@@ -9,8 +9,10 @@ export const NS_MODULE = 'modules';
 
 export const K_ENV = 'env';
 
-const runtime = new RuntimeRegistry();
+export const runtime = new RuntimeRegistry();
 runtime.register('runtimePath', __dirname);
 runtime.loadModule = (name, m) => runtime.register(NS_MODULE, name, m);
+
+export const Runtime = runtime;
 
 export default runtime;