+++ packages/jacaranda/src/starters/index.js
@@ -2,3 +2,4 @@ export { default as startWorker } from './worker';
 export { default as startLoopWorker } from './loopWorker';
 export { default as startCommand } from './command';
 export { default as startScheduler, cronPatterns } from './schedulerWorker';
+export { default as startQueueWorker } from './queueWorker';