+++ packages/tester/src/asyncDump.js
@@ -2,6 +2,7 @@ import { createHook } from 'node:async_hooks';
 import { stackTraceFilter } from 'mocha/lib/utils';
 import path from 'node:path';
 import fs from 'node:fs/promises';
+
 const allResources = new Map();
 
 // this will pull Mocha internals out of the stacks